refactor: 非shadcn/uiコンポーネントをcomponents/からapp/components/に移動する (#1023)
components/ディレクトリの配置基準を明確化し、shadcn/ui生成物のみを
配置する方針に統一する。error-content.tsxとcalendar/を移動し、
参照元のimportパスを更新。
