+++ app/(authenticated)/circles/components/circle-overview-calendar.tsx
@@ -3,7 +3,7 @@
 import {
   SessionCalendar,
   type SessionExtendedProps,
-} from "@/components/calendar/session-calendar";
+} from "@/app/components/calendar/session-calendar";
 import { toLocalDateString } from "@/lib/date-utils";
 import { Button } from "@/components/ui/button";
 import type { CircleOverviewSession } from "@/server/presentation/view-models/circle-overview";
