+++ app/components/calendar/session-calendar.css
